% Auto-generated: do not edit by hand
\name{dashInstaStories}

\alias{dashInstaStories}

\title{DashInstaStories component}

\description{

}

\usage{
dashInstaStories(id=NULL, currentIndex=NULL, defaultInterval=NULL,
header=NULL, height=NULL, isPaused=NULL,
keyboardNavigation=NULL, loader=NULL, loop=NULL,
onAllStoriesEnd=NULL, onNext=NULL, onPrevious=NULL,
onStoryEnd=NULL, onStoryStart=NULL, preloadCount=NULL,
preventDefault=NULL, progressContainerStyles=NULL,
progressStyles=NULL, progressWrapperStyles=NULL,
renderers=NULL, stories=NULL, storyContainerStyles=NULL,
storyStyles=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{currentIndex}{Numeric. }

\item{defaultInterval}{Numeric. }

\item{header}{}

\item{height}{Numeric | character. }

\item{isPaused}{Logical. }

\item{keyboardNavigation}{Logical. }

\item{loader}{}

\item{loop}{Logical. }

\item{onAllStoriesEnd}{}

\item{onNext}{}

\item{onPrevious}{}

\item{onStoryEnd}{}

\item{onStoryStart}{}

\item{preloadCount}{Numeric. }

\item{preventDefault}{Logical. }

\item{progressContainerStyles}{Named list. }

\item{progressStyles}{Named list. }

\item{progressWrapperStyles}{Named list. }

\item{renderers}{Unnamed list. }

\item{stories}{List of character | lists containing elements 'url', 'header'.
those elements have the following types:
  - url (character; required)
  - header (optional): . header has the following type: lists containing elements 'heading', 'subheading', 'profileimage'.
those elements have the following types:
  - heading (character; optional)
  - subheading (character; optional)
  - profileimage (character; optional)s. }

\item{storyContainerStyles}{Named list. }

\item{storyStyles}{Named list. }

\item{width}{Numeric | character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

